# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API de Índice de Masa Corporal",
    "description": "Calcula el Índice de Masa Corporal (IMC = peso / estatura²) y lo clasifica según las categorías de la Organización Mundial de la Salud.",
    "contact": {
      "name": "Curso de DevOps - Platzi",
      "url": "https://github.com/SerAle1105/curso-devops-platzi"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Bmi": {
      "post": {
        "tags": [
          "IMC"
        ],
        "summary": "Calcula el Índice de Masa Corporal a partir del peso y la estatura.",
        "description": "Petición de ejemplo:\r\n            \r\n    POST /api/bmi\r\n    {\r\n       \"weightKg\": 70,\r\n       \"heightM\": 1.75\r\n    }\r\n            \r\nEl IMC se calcula como peso / estatura² y se redondea a dos decimales.",
        "requestBody": {
          "description": "Peso en kilogramos y estatura en metros.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BmiRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BmiRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BmiRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IMC calculado con su clasificación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BmiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Los datos enviados no son válidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "IMC"
        ],
        "summary": "Calcula el IMC por query string.",
        "description": "Variante pensada para pruebas rápidas desde el navegador o desde Swagger:\r\n            \r\n    GET /api/bmi?weightKg=70&heightM=1.75",
        "parameters": [
          {
            "name": "weightKg",
            "in": "query",
            "description": "Peso en kilogramos. Rango admitido: 1 a 500.",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "heightM",
            "in": "query",
            "description": "Estatura en metros. Rango admitido: 0.5 a 2.5.",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "IMC calculado con su clasificación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BmiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Los datos enviados no son válidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BmiRequest": {
        "required": [
          "heightM",
          "weightKg"
        ],
        "type": "object",
        "properties": {
          "weightKg": {
            "maximum": 500,
            "minimum": 1,
            "type": "number",
            "description": "Peso de la persona en kilogramos.",
            "format": "double",
            "example": 70
          },
          "heightM": {
            "maximum": 2.5,
            "minimum": 0.5,
            "type": "number",
            "description": "Estatura de la persona en metros.",
            "format": "double",
            "example": 1.75
          }
        },
        "additionalProperties": false,
        "description": "Datos de entrada para el cálculo del Índice de Masa Corporal."
      },
      "BmiResponse": {
        "type": "object",
        "properties": {
          "weightKg": {
            "type": "number",
            "description": "Peso recibido, en kilogramos.",
            "format": "double",
            "example": 70
          },
          "heightM": {
            "type": "number",
            "description": "Estatura recibida, en metros.",
            "format": "double",
            "example": 1.75
          },
          "bmi": {
            "type": "number",
            "description": "IMC calculado, redondeado a dos decimales.",
            "format": "double",
            "example": 22.86
          },
          "category": {
            "type": "string",
            "description": "Clasificación del IMC según la OMS.",
            "nullable": true,
            "example": "Peso normal"
          }
        },
        "additionalProperties": false,
        "description": "Resultado del cálculo del Índice de Masa Corporal."
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      }
    }
  },
  "tags": [
    {
      "name": "IMC",
      "description": "Cálculo del Índice de Masa Corporal."
    }
  ]
}